Audit fixes update

+++ contracts/interfaces/ISafeMinimalMultisig.sol
@@ -9,15 +9,15 @@ import "./ICryptoLegacyBuildManager.sol";
 import "@openzeppelin/contracts/utils/structs/EnumerableSet.sol";
 
 interface ISafeMinimalMultisig {
-    event CreateSafeMinimalMultisigProposal(uint256 proposalId, bytes32 voter, uint8 reqConfirmations);
-    event CancelSafeMinimalMultisigProposal(uint256 proposalId, bytes32 voter, uint8 reqConfirmations, ProposalStatus status);
-    event ConfirmSafeMinimalMultisigProposal(uint256 proposalId, bytes32 voter, uint8 reqConfirmations, uint256 confirms);
+    event CreateSafeMinimalMultisigProposal(uint256 proposalId, bytes32 voter, uint128 reqConfirmations);
+    event CancelSafeMinimalMultisigProposal(uint256 proposalId, bytes32 voter, uint128 reqConfirmations, ProposalStatus status);
+    event ConfirmSafeMinimalMultisigProposal(uint256 proposalId, bytes32 voter, uint128 reqConfirmations, uint256 confirms);
     event ExecuteSafeMinimalMultisigProposal(uint256 proposalId, bytes32 voter, bool executed, bytes returnData);
-    event SetVotersAndConfirmations(bytes32[] voters, uint8 requiredConfirmations);
-    event SetConfirmations(uint8 requiredConfirmations);
+    event SetVotersAndConfirmations(bytes32[] voters, uint128 requiredConfirmations);
+    event SetConfirmations(uint128 requiredConfirmations);
 
     struct Storage {
-        uint8 requiredConfirmations;
+        uint128 requiredConfirmations;
         bytes32[] voters;
         Proposal[] proposals;
         mapping(uint256 => mapping(bytes32 => bool)) confirmedBy;
@@ -32,7 +32,7 @@ interface ISafeMinimalMultisig {
 
     struct Proposal {
         bytes params;
-        uint8 confirms;
+        uint128 confirms;
         bytes4 selector;
         ProposalStatus status;
     }
